+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### fix
 
 * add more tracking for matomo [#4129](https://github.com/MyEtherWallet/MyEtherWallet/pull/4129)
+* add ledger debug warning [#4141](https://github.com/MyEtherWallet/MyEtherWallet/pull/4141)
 
 ### Release v6.6.4